clean up lib directory by moving testing code to its own file lib_test.rs
